requirements.txt
amqp==5.3.1
arabic-reshaper==3.0.0
asgiref==3.11.0
asn1crypto==1.5.1
billiard==4.2.4
celery==5.4.0
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
cryptography==46.0.3
cssselect2==0.8.0
Deprecated==1.3.1
Django==5.2.8
djangorestframework==3.16.1
freetype-py==2.5.1
gunicorn==23.0.0
html5lib==1.1
idna==3.11
kombu==5.6.1
lxml==6.0.2
oscrypto==1.3.0
packaging==25.0
pikepdf==10.0.2
pillow==12.0.0
prompt_toolkit==3.0.52
psycopg2-binary==2.9.11
pycairo==1.29.0
pycparser==2.23
pyHanko==0.32.0
pyhanko-certvalidator==0.29.0
pypdf==6.4.0
python-bidi==0.6.7
python-dateutil==2.9.0.post0
PyYAML==6.0.3
redis==5.0.0
reportlab==4.4.5
requests==2.32.5
rlPyCairo==0.4.0
six==1.17.0
sqlparse==0.5.4
svglib==1.6.0
tinycss2==1.5.1
tzdata==2025.2
tzlocal==5.3.1
uritools==5.0.0
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.14
webencodings==0.5.1
whitenoise==6.11.0
wrapt==2.0.1
xhtml2pdf==0.2.17
